requirements.txt
mysql==0.0.3
mysql-connector-python==8.0.33
gspread==5.10.0
pandas==2.0.2
oauth2client==4.1.3
google-api-python-client==2.91.0
requests==2.31.0
google-cloud-bigquery==2.28.0
pyarrow==6.0.1